IndexingHelpers now broken into 3 different classes for clarity on responsibilities
